pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "waveform"
version = "2.0.0.dev0"
description = "AI-powered event playlist designer for Spotify"
requires-python = ">=3.10"
dependencies = [
    "spotipy>=2.23.0",
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]
audio-pygame = ["pygame>=2.5.0"]
audio-vlc = ["python-vlc>=3.0.0"]
analytics = ["posthog>=3.0.0"]
ui = ["customtkinter>=5.2.0"]

[project.scripts]
waveform = "waveform.app.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["waveform*"]
exclude = ["waveform/tests*"]

[tool.pytest.ini_options]
testpaths = ["waveform/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # long lines are fine in prompt strings
